dom/locks/test/test_strongworker.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css"/>

<script>
  SimpleTest.waitForExplicitFinish();

  async function run() {
    // Bug 1746646: Make mochitests work with TCP enabled (cookieBehavior = 5)
    // Acquire storage access permission here so that the BroadcastChannel used to
    // communicate with the opened windows works in xorigin tests. Otherwise,
    // the iframe containing this page is isolated from first-party storage access,
    // which isolates BroadcastChannel communication.
    if (isXOrigin) {
      SpecialPowers.wrap(document).notifyUserGestureActivation();
      await SpecialPowers.addPermission(
        "storageAccessAPI",
        true,
        window.location.href
       );
      await SpecialPowers.wrap(document).requestStorageAccess();
    }
    const channel = new BroadcastChannel("strongworker");
    await navigator.locks.request("exclusive", async () => {
      await new Promise(resolve => {
        let worker = new Worker("./file_strongworker.js");
        worker.onmessage = resolve; // onload
      });
      const query = await navigator.locks.query();
      is(query.pending.length, 1, "Pending request exists");

      // Garbage collect the worker
      SpecialPowers.DOMWindowUtils.garbageCollect();
    });

    channel.onmessage = async event => {
      const query = await navigator.locks.query();
      is(query.pending.length, 0, "No pending request");
      SimpleTest.finish();
    };
  }
  run();
</script>
